requirements.txt
streamlit
pdf2docx
Pillow
PyPDF2
PyMuPDF
streamlit-sortables
# pydub
